refactor: abstract "toFormattedDataForProps" to export
