+++ src/plugin/reactPlugin/index.js
@@ -7,42 +7,7 @@ const addShortcode = require('./addShortcode')
 const { stringify } = require('javascript-stringify')
 const toRendererHtml = require('./toRendererHtml')
 const htmlEscape = require('../../utils/htmlEscape')
-
-const SLINKITY_REACT_RENDERER_PATH =
-  'slinkity/lib/plugin/reactPlugin/_slinkity-react-renderer.js'
-function formatDataForProps(eleventyData = {}) {
-  const formattedCollections = {}
-  for (let name of Object.keys(eleventyData.collections)) {
-    formattedCollections[name] = eleventyData.collections[name].map((item) => {
-      /**
-       * Items omitted:
-       * - template: reference to the Template class
-       *   - reason: this is an unwieldy object that's nearly impossible to stringify for clientside use
-       * - templateContent setter: method to set rendered templateContent
-       *   - reason: components shouldn't be able to set underlying 11ty data
-       * - data.collections: a nested, circular reference to the outer collections
-       *   - reason: it's best to remove circular dependencies before stringifying for clientside use
-       */
-      const { collections, ...data } = item.data
-      return {
-        inputPath: item.inputPath,
-        fileSlug: item.fileSlug,
-        filePathStem: item.filePathStem,
-        date: item.date,
-        outputPath: item.outputPath,
-        url: item.url,
-        data,
-        get templateContent() {
-          return item.templateContent
-        },
-      }
-    })
-  }
-  return {
-    ...eleventyData,
-    collections: formattedCollections,
-  }
-}
+const toFormattedDataForProps = require('./toFormattedDataForProps')
 
 module.exports = function reactPlugin(eleventyConfig, { dir }) {
   eleventyConfig.addTemplateFormats('jsx')
@@ -68,7 +33,7 @@ module.exports = function reactPlugin(eleventyConfig, { dir }) {
         const { default: Component = () => null, getProps = () => ({}) } =
           await toCommonJSModule({ inputPath })
 
-        const props = getProps(formatDataForProps(data))
+        const props = getProps(toFormattedDataForProps(data))
         componentToPropsMap[jsxImportPath] = props
 
         return toRendererHtml({
